refac: split server file, use axios instead of fetch
