+++ src/utils/convert-ft-balance.ts
@@ -1,0 +1,3 @@
+export function convertFTBalance(value: string, decimals: number) {
+  return (parseFloat(value) / Math.pow(10, decimals)).toFixed(2);
+}
